// backend/src/core/enums/files.ts
export enum FileType {
  AVATAR = 'avatar',
  COVER = 'cover',
  MUSICPREVIEW = 'musicpreview',
  MUSIC = 'music',
}

export enum BucketType {
  AVATAR = 'AWS_AVATAR_BUCKET',
  PREVIEW = 'AWS_PREVIEW_BUCKET',
  MUSIC = 'AWS_MUSIC_BUCKET',
  MUSIC_COMPRESSED = 'AWS_MUSIC_COMPRESSED_BUCKET',
}

export enum MusicExtension {
  MP3 = 'mp3',
  WAV = 'wav',
  STEMS = 'stems',
}